customer_management/requirements.txt
Django==3.2.8
djangorestframework==3.14.0
openpyxl==3.1.2
pandas==2.1.4